+++ Sources/VComponents/Components/Indicators/Page Indicator/Compact/VPageIndicatorCompactUIModel.swift
@@ -1,5 +1,5 @@
 //
-//  VPageIndicatorInfiniteUIModel.swift
+//  VPageIndicatorCompactUIModel.swift
 //  VComponents
 //
 //  Created by Vakhtang Kontridze on 25.02.23.
@@ -8,11 +8,11 @@
 import SwiftUI
 import VCore
 
-// MARK: - V Page Indicator Infinite UI Model
+// MARK: - V Page Indicator Compact UI Model
 /// Model that describes UI.
-public struct VPageIndicatorInfiniteUIModel {
+public struct VPageIndicatorCompactUIModel {
     // MARK: Properties
-    fileprivate static let pageIndicatorFiniteReference: VPageIndicatorFiniteUIModel = .init()
+    fileprivate static let pageIndicatorStandardReference: VPageIndicatorStandardUIModel = .init()
     
     /// Sub-model containing layout properties.
     public var layout: Layout = .init()
@@ -32,7 +32,7 @@ public struct VPageIndicatorInfiniteUIModel {
     public struct Layout {
         // MARK: Properties
         /// Direction. Defaults to `leftToRight`.
-        public var direction: OmniLayoutDirection = pageIndicatorFiniteReference.layout.direction
+        public var direction: OmniLayoutDirection = pageIndicatorStandardReference.layout.direction
         
         /// Number of visible dots. Default to `7`.
         ///
@@ -49,18 +49,18 @@ public struct VPageIndicatorInfiniteUIModel {
         var middleDots: Int { visibleDots / 2 }
         
         /// Dot dimension. Defaults to `10`.
-        public var dotDimension: CGFloat = pageIndicatorFiniteReference.layout.dotDimension
+        public var dotDimension: CGFloat = pageIndicatorStandardReference.layout.dotDimension
 
         /// Scale of dot at the edge. Defaults to `0.5`.
         ///
         /// If there are `7` visible dots, and `3` center dots, scales would sit at `[0.5, 0.75, 1, 1, 1, 0.75, 0.5]`.
         public var edgeDotScale: CGFloat = 0.5
         
-        /// Unselected dot scale for when component turns to `finite` configuration. Defaults to `0.85`.
-        public var unselectedDotScaleForFiniteConfiguration: CGFloat = pageIndicatorFiniteReference.layout.unselectedDotScale
+        /// Unselected dot scale when switching to `standard` configuration. Defaults to `0.85`.
+        public var unselectedDotScaleForStandardConfiguration: CGFloat = pageIndicatorStandardReference.layout.unselectedDotScale
         
         /// Dot spacing. Defaults to `5`.
-        public var spacing: CGFloat = pageIndicatorFiniteReference.layout.spacing
+        public var spacing: CGFloat = pageIndicatorStandardReference.layout.spacing
         
         // MARK: Initializers
         /// Initializes sub-model with default values.
@@ -69,19 +69,19 @@ public struct VPageIndicatorInfiniteUIModel {
     
     // MARK: Colors
     /// Sub-model containing color properties.
-    public typealias Colors = VPageIndicatorFiniteUIModel.Colors
+    public typealias Colors = VPageIndicatorStandardUIModel.Colors
 
     // MARK: Animations
     /// Sub-model containing animation properties.
-    public typealias Animations = VPageIndicatorFiniteUIModel.Animations
+    public typealias Animations = VPageIndicatorStandardUIModel.Animations
     
     // MARK: Sub-Models
-    var finiteSubModel: VPageIndicatorFiniteUIModel {
-        var uiModel: VPageIndicatorFiniteUIModel = .init()
+    var standardSubModel: VPageIndicatorStandardUIModel {
+        var uiModel: VPageIndicatorStandardUIModel = .init()
         
         uiModel.layout.direction = layout.direction
         uiModel.layout.dotDimension = layout.dotDimension
-        uiModel.layout.unselectedDotScale = layout.unselectedDotScaleForFiniteConfiguration
+        uiModel.layout.unselectedDotScale = layout.unselectedDotScaleForStandardConfiguration
         uiModel.layout.spacing = layout.spacing
         
         uiModel.colors = colors