+++ Sources/VComponents/Deprecations.swift
@@ -29,18 +29,18 @@ extension VPageIndicator {
 public typealias VPageIndicatorUIModel = VPageIndicatorAutomaticUIModel
 
 extension VPageIndicatorType {
-    @available(*, deprecated, message: "Use method with UI model instead")
+    @available(*, deprecated, message: "Use method with UI model instead. Also, renamed to `standard`.")
     public static var finite: Self {
-        finite()
+        standard()
     }
     
-    @available(*, deprecated, message: "Use method with UI model instead")
+    @available(*, deprecated, message: "Use method with UI model instead. Also, renamed to `compact`.")
     public static func infinite(
         visible: Int = 7,
         center: Int = 3
     ) -> Self {
-        infinite(uiModel: {
-            var uiModel: VPageIndicatorInfiniteUIModel = .init()
+        compact(uiModel: {
+            var uiModel: VPageIndicatorCompactUIModel = .init()
             uiModel.layout.visibleDots = visible
             uiModel.layout.centerDots = center
             return uiModel
@@ -57,9 +57,23 @@ extension VPageIndicatorType {
             var uiModel: VPageIndicatorAutomaticUIModel = .init()
             uiModel.layout.visibleDots = visible
             uiModel.layout.centerDots = center
-            uiModel.layout.finiteDotLimit = finiteLimit
+            uiModel.layout.compactDotLimit = finiteLimit
             return uiModel
         }())
+    }
+}
+
+@available(*, deprecated, renamed: "VPageIndicatorStandardUIModel")
+public typealias VPageIndicatorFiniteUIModel = VPageIndicatorStandardUIModel
+
+@available(*, deprecated, renamed: "VPageIndicatorCompactUIModel")
+public typealias VPageIndicatorInfiniteUIModel = VPageIndicatorCompactUIModel
+
+extension VPageIndicatorAutomaticUIModel.Layout {
+    @available(*, deprecated, renamed: "compactDotLimit")
+    public var finiteDotLimit: Int {
+        get { compactDotLimit }
+        set { compactDotLimit = newValue }
     }
 }